// src/app/core/models/record.dto.ts
export interface RecordDto {
    id?: number; // autogenerated
    title: string;
    artist: string;
    format: string;
    genre: string;
    releaseYear: number;
    price: number;
    stockQuantity: number;

    //customer fields
    customerId: string;
    customerFirstName: string;
    customerLastName: string;
    customerContactNumber: string;
    customerEmail: string;
}